# td-level-editor/td-level-editor-http/src/main/resources/openapi/td-level-studio-api.yaml
openapi: 3.0.3
info:
  title: Tower Defense Level Studio API
  description: |
    API complète pour l'éditeur de niveaux Tower Defense.
    
    Cette API permet de :
    - Créer et éditer des niveaux
    - Gérer les types d'ennemis
    - Gérer les types de tours
    - Importer/exporter en format draft (éditable) ou exportable (pour le moteur)
  version: 2.0.0
  contact:
    name: Tower Defense Studio Team
servers:
  - url: http://localhost:8081
    description: Serveur de développement local
...
paths:
...
  /api/levels/{levelId}/metadata:
    patch:
      summary: Met à jour les métadonnées d'un niveau
      tags: [Level Editing]
      operationId: updateLevelMetadata
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LevelMetadataUpdate'
      responses:
        '200':
          description: Métadonnées mises à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditableLevel'
  /api/levels/{levelId}/paths:
    get:
      summary: Liste tous les chemins d'un niveau
      tags: [Level Editing]
      operationId: getLevelPaths
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liste des chemins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditablePath'
    
    post:
      summary: Ajoute un nouveau chemin au niveau
      tags: [Level Editing]
      operationId: addPath
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditablePath'
      responses:
        '201':
          description: Chemin ajouté
  /api/levels/{levelId}/paths/{pathId}:
    delete:
      summary: Supprime un chemin du niveau
      tags: [Level Editing]
      operationId: removePath
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
        - name: pathId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Chemin supprimé
  /api/levels/{levelId}/attacks:
    get:
      summary: Liste toutes les attaques d'un niveau
      tags: [Level Editing]
      operationId: getLevelAttacks
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liste des attaques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditableAttack'
    
    post:
      summary: Ajoute une nouvelle attaque au niveau
      tags: [Level Editing]
      operationId: addAttack
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttackRequest'
      responses:
        '201':
          description: Attaque ajoutée
  /api/levels/{levelId}/attacks/{attackId}:
    delete:
      summary: Supprime une attaque du niveau
      tags: [Level Editing]
      operationId: removeAttack
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
        - name: attackId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Attaque supprimée
  /api/levels/{levelId}/attacks/{attackId}/waves:
    post:
      summary: Ajoute une vague à une attaque
      tags: [Level Editing]
      operationId: addWave
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
        - name: attackId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableWave'
      responses:
        '201':
          description: Vague ajoutée
  /api/levels/{levelId}/attacks/{attackId}/waves/{waveId}:
    delete:
      summary: Supprime une vague d'une attaque
      tags: [Level Editing]
      operationId: removeWave
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
        - name: attackId
          in: path
          required: true
          schema:
            type: string
        - name: waveId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Vague supprimée
  /api/levels/{levelId}/tower-types:
    post:
      summary: Ajoute un type de tour disponible au niveau
      tags: [Level Editing]
      operationId: addTowerTypeToLevel
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTowerTypeRequest'
      responses:
        '204':
          description: Type de tour ajouté
  /api/levels/{levelId}/tower-types/{towerTypeId}:
    delete:
      summary: Retire un type de tour du niveau
      tags: [Level Editing]
      operationId: removeTowerTypeFromLevel
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Type de tour retiré

  # ========================================================================
  # ENEMIES COLLECTION
  # ========================================================================
  /api/enemies:
    get:
      summary: Liste tous les types d'ennemis
      tags: [Enemies]
      operationId: getAllEnemyTypes
      responses:
        '200':
          description: Liste des types d'ennemis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditableEnemyType'
    
    post:
      summary: Crée un nouveau type d'ennemi
      tags: [Enemies]
      operationId: createEnemyType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnemyTypeRequest'
      responses:
        '201':
          description: Type d'ennemi créé
  /api/enemies/{enemyTypeId}:
    get:
      summary: Récupère un type d'ennemi par son ID
      tags: [Enemies]
      operationId: getEnemyType
      parameters:
        - name: enemyTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Détails du type d'ennemi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditableEnemyType'
    
    delete:
      summary: Supprime un type d'ennemi
      tags: [Enemies]
      operationId: deleteEnemyType
      parameters:
        - name: enemyTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Type d'ennemi supprimé
  /api/enemies/import/draft:
    post:
      summary: Importe un type d'ennemi depuis un fichier draft
      tags: [Enemies]
      operationId: importDraftEnemyType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201':
          description: Type d'ennemi importé
  /api/enemies/import/exportable:
    post:
      summary: Importe un type d'ennemi depuis un fichier exportable
      tags: [Enemies]
      operationId: importExportableEnemyType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201':
          description: Type d'ennemi importé
  /api/enemies/{enemyTypeId}/export/draft:
    post:
      summary: Exporte un type d'ennemi en format draft
      tags: [Enemies]
      operationId: exportDraftEnemyType
      parameters:
        - name: enemyTypeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Type d'ennemi exporté
  /api/enemies/{enemyTypeId}/export/exportable:
    post:
      summary: Exporte un type d'ennemi en format exportable
      tags: [Enemies]
      operationId: exportExportableEnemyType
      parameters:
        - name: enemyTypeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Type d'ennemi exporté

  # ========================================================================
  # TOWERS COLLECTION
  # ========================================================================
...
components:
  schemas:
...
    ImportRequest:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          description: Chemin du fichier à importer
    ExportRequest:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          description: Chemin où sauvegarder le fichier
    LevelMetadataUpdate:
      type: object
      properties:
        startingMoney:
          type: integer
          minimum: 0
        startingLives:
          type: integer
          minimum: 1
    CreateAttackRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    AddTowerTypeRequest:
      type: object
      required:
        - towerTypeId
      properties:
        towerTypeId:
          type: string
    CreateEnemyTypeRequest:
      type: object
      required:
        - id
        - health
        - speed
        - reward
      properties:
        id:
          type: string
        health:
          type: integer
          minimum: 1
        speed:
          type: number
          format: double
          minimum: 0.01
        reward:
          type: integer
          minimum: 0
...
    EditableLevel:
      type: object
      required:
        - metadata
        - map
        - paths
        - attacks
        - towerCapacityIds
      properties:
        metadata:
          $ref: '#/components/schemas/LevelMetadata'
        map:
          $ref: '#/components/schemas/EditableMap'
        paths:
          type: array
          items:
            $ref: '#/components/schemas/EditablePath'
        attacks:
          type: array
          items:
            $ref: '#/components/schemas/EditableAttack'
        towerCapacityIds:
          type: array
          items:
            type: string
    LevelMetadata:
      type: object
      required:
        - id
        - startingMoney
        - startingLives
      properties:
        id:
          type: string
        startingMoney:
          type: integer
          minimum: 0
        startingLives:
          type: integer
          minimum: 1
    EditableMap:
      type: object
      required:
        - width
        - height
        - paths
      properties:
        width:
          type: integer
        height:
          type: integer
        paths:
          type: array
          items:
            $ref: '#/components/schemas/EditablePath'
    EditablePath:
      type: object
      required:
        - id
        - points
      properties:
        id:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/PositionDefinition'
          minItems: 2
    PositionDefinition:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
        y:
          type: integer
    EditableAttack:
      type: object
      required:
        - id
        - waves
      properties:
        id:
          type: string
        waves:
          type: array
          items:
            $ref: '#/components/schemas/EditableWave'
    EditableWave:
      type: object
      required:
        - id
        - startTick
        - spawnInterval
        - count
        - enemyType
        - pathId
      properties:
        id:
          type: string
        startTick:
          type: integer
          minimum: 0
        spawnInterval:
          type: integer
          minimum: 1
        count:
          type: integer
          minimum: 1
        enemyType:
          type: string
        pathId:
          type: string
    EditableEnemyType:
      type: object
      required:
        - id
        - health
        - speed
        - reward
      properties:
        id:
          type: string
        health:
          type: integer
          minimum: 1
        speed:
          type: number
          format: double
          minimum: 0.01
        reward:
          type: integer
          minimum: 0
...